refactor:#8 pr feedback
