+++ src/app/Router.jsx
@@ -2,7 +2,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import HomePage from '@pages/HomePage';
 import ViewPage from '@pages/view/ViewPage';
 import SigninPage from '@pages/sign-in/SignInPage';
-import RootLayout from '@/components/layouts/RootLayout';
+import RootLayout from '@layouts/RootLayout';
 
 const publicRoutes = [
   {
@@ -14,14 +14,14 @@ const publicRoutes = [
         element: <HomePage />
       },
       {
-        path: `/code/view/:id`,
+        path: '/code/view/:id',
         element: <ViewPage />
-      },
-      {
-        path: '/signin',
-        element: <SigninPage />
       }
     ]
+  },
+  {
+    path: '/signin',
+    element: <SigninPage />
   }
 ];
