+++ src/pages/view/ViewPage.jsx
@@ -4,7 +4,9 @@ import PostAreaContainer from '@features/view-page/PostAreaContainer';
 import CommentAreaContainer from '@features/view-page/CommentAreaContainer';
 
 export default function ViewPage() {
-  const PostWriter = {
+  const postProps = {
+    // 페이지 본문의 동적 변환 값을 props를 통해 임시적으로 표현하기 위한 객체
+    // 추후 이 부분은 supabase의 db로 대체될 예정
     IsAuth: true,
     Title: '제목',
     CodeLanguage: 'Language',
@@ -14,7 +16,9 @@ export default function ViewPage() {
     Likes: 0
   };
 
-  const CommentWriter = {
+  const commentProps = {
+    // 페이지 댓글의 동적 변환 값을 props를 통해 임시적으로 표현하기 위한 객체
+    // 추후 이 부분은 supabase의 db로 대체될 예정
     IsAuth: true,
     Nickname: '댓글 작성자',
     Contents: '댓글입니다.',
@@ -23,8 +27,10 @@ export default function ViewPage() {
 
   return (
     <StViewPageContainer>
-      <PostAreaContainer PostWriter={PostWriter} />
-      <CommentAreaContainer CommentWriter={CommentWriter} />
+      <PostAreaContainer postProps={postProps} />{' '}
+      {/* 페이지의 본문에 DB에 저장되어 있는 값에 따라 변화하는 값을 임의적으로 표현하기 위해 내린 props */}
+      <CommentAreaContainer commentProps={commentProps} />{' '}
+      {/* 페이지 댓글 상자에 DB에 저장되어 있는 값에 따라 변화하는 값을 임의적으로 표현하기 위해 내린 props */}
     </StViewPageContainer>
   );
 }
